refactor(ui): extract bounded context row render data
